Time: 0 ms (100%), Space: 63 MB (49.49%) - LeetHub

+++ 0011-container-with-most-water/0011-container-with-most-water.cpp
@@ -1,13 +1,13 @@
 class Solution {
 public:
     int maxArea(vector<int>& height) {
-        long long minheight = 0;
-        long long maxarea = 0;
+        int  minheight = 0;
+        int maxarea = 0;
         int i = 0;
         int j = height.size()- 1;
         while(i <= j){
             minheight = min(height[i], height[j]);
-            long long area = minheight * (j - i);
+            int area = minheight * (j - i);
             maxarea = max(maxarea, area);
             if(height[i] < height[j]){
                 i++;
